[VEUE-408] Broadcaster: Logout and Menu Cleanup
In this PR we do several things:

* Change menu.js to Typescript
* Remove unused or unwanted menu items
* Show Debug menu in all builds (for now)
* ALLOW FOR YOU TO LOG OUT!?!?!?!!!!!!!!!!!! YAY!
